Refactor for organization
