+++ unlock-app/src/components/interface/VerificationStatus.tsx
@@ -2,13 +2,13 @@ import React from 'react'
 import sigUtil from 'eth-sig-util'
 import { useQuery } from '@apollo/react-hooks'
 import { ApolloError } from 'apollo-boost'
-import { ethers } from 'ethers'
 import { DefaultError } from '../creator/FatalError'
 import {
   durationsAsTextFromSeconds,
   expirationAsDate,
 } from '../../utils/durations'
 import { OwnedKey } from './keyChain/KeychainTypes'
+import RefundButton from './RefundButton'
 import keyHolderQuery from '../../queries/keyHolder'
 import withConfig from '../../utils/withConfig'
 import 'cross-fetch/polyfill'
@@ -100,61 +100,6 @@ export const VerificationStatus = ({ data, sig, hexData, config }: Props) => {
         )}
     </div>
   )
-}
-
-interface RefundButtonProps {
-  externalRefundContractAddress: string
-  accountAddress: string
-  provider: any
-}
-interface RefundButtonState {
-  refundInitiated: boolean
-}
-export class RefundButton extends React.Component<
-  RefundButtonProps,
-  RefundButtonState
-> {
-  private contract: ethers.Contract
-  constructor(props: RefundButtonProps) {
-    super(props)
-    const { provider, externalRefundContractAddress } = props
-    const web3Provider = new ethers.providers.Web3Provider(provider)
-    const signer = web3Provider.getSigner()
-    const abi = ['function refund(address recipient)']
-    this.contract = new ethers.Contract(
-      externalRefundContractAddress,
-      abi,
-      signer
-    )
-
-    this.state = {
-      refundInitiated: false,
-    }
-  }
-
-  initiateRefund = async () => {
-    const { accountAddress } = this.props
-    await this.contract.refund(accountAddress)
-    this.setState({
-      refundInitiated: true,
-    })
-  }
-
-  render() {
-    const { refundInitiated } = this.state
-    if (refundInitiated) {
-      return (
-        <button type="button" disabled>
-          Refund Initiated
-        </button>
-      )
-    }
-    return (
-      <button type="button" onClick={this.initiateRefund}>
-        Perform refund
-      </button>
-    )
-  }
 }
 
 export const Identity = ({ valid }: { valid: boolean }) => (
